Finalize repo name change. fix bypassess, change around script container
Fixed bypasses which allowed code to execute outside it, also fixed a potential issue with specifc code not working properly due to lack of support in the wrapper for tuple returns.

Changed the .rbxl file name to reflecrt the new repo name - and updated .gitignore to reflect this too.

Changed around the script container for running server code - now just executes the function, rather than setfenving the function and using the function returned from that.

Also minor change: added a way for the client to retrieve specific configuration settings from the server. Currently only PLACE_NAME is necessary.

The console GUI is now able to reflect the place's name.