Re-wrote sandbox entirely, temporarily prevent require from being called
Re-wrote the entirety of the sandbox. Now only wraps and sandboxes what needs to be sandboxed - more work is needed however.

The following functions still need to be sandboxed:
- type
- typeof

Errors are now no longer "faked" - since there really isn't a need to fake them anymore.

When scripts are ran it now sends feedback to the console to indicate execution.