Update to work with new backend, revise LocalScript execution process
The new backend supports using modules instead of just local scripts for local script execution - which allows people to just require the module without the need for a custom backend and/or go through the process to set up their own.

This also adds a note to the top of some the code to signify license and source location information.

The option for using non-module scripts still is there and works, the current backend for this project however just doesn't support it natively (yet). The option currently requires a custom backend, but in the future will change once re-implemented in the currrent code for the backend for this project.