Slight cleanup, updates
Slight clean up of the repository, deleting the Archive directory as everything inside it is no longer of relevance.

Add some linter settings to vscode, so that way there is nothing in the problems section of vscode anymore.

Adds support for when shared is already locked, such as in other script builders.

Modifies the sandbox for specific edge cases where unwrapped values are needed in order to be passed to functions. Adds method Sandbox.wrapFunction() to support this.

Also fixes an issue with a specific edge case of realArgs inside Sandbox.wrapMethod() being nil causing an error in the sandbox.

Also some very very small cleanup is done in the sandbox too.

The Script Builder Project module has been modified in order to better handle output in specific circumstances, and should no longer error when the backend fails to initialize the script builder, allowing the   script builder to still run. Although in the future this might not be needed, a warning might be added to let the user knoiw about this.

Minor change to README to reflect current repo URL for the backend.