Fix critical issue in Sandbox.lua
This fixes an issue with the Sandbox which caused issues with tuples being returned from functions. When a function is called and returns a tuple, it can cause data returned to be lost. This is fixed by wrapping the returned data inside a table and then unpacking the result, after determining a success or failure inside the pcall.